speedrun/Web/paper-bouqet/src/src/backend/requirements.txt
django
pyyaml